requirements.txt
streamlit>=1.25.0
opencv-python>=4.7.0
numpy>=1.25.0
pandas>=2.1.0
plotly>=5.18.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.42.0
pillow>=10.0.0
pytesseract>=0.3.10